+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fixed an "Element query executed before Craft is fully initialized" warning in the Craft logs that was caused by Content Templates
+
 ## 1.0.2 - 2023-08-23
 
 ### Fixed