+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.0.3 - 2023-08-25
 
 ### Fixed
 - Fixed a bug where the 'Choose a template' modal would not appear when creating a new entry with fields with non-empty default values